Initiatives: extract EnableSwitch
